Move functions to config_test.go to isolate tags
