+++ vcd/config_test.go
@@ -1184,3 +1184,67 @@ func skipNoNsxtConfiguration(t *testing.T) {
 		t.Skip(generalMessage + "No VRF NSX-T Tier-0 specified")
 	}
 }
+
+func testAccCheckVcdStandaloneVmExists(vmName, node, orgName, vdcName string) resource.TestCheckFunc {
+	if orgName == "" {
+		orgName = testConfig.VCD.Org
+	}
+	if vdcName == "" {
+		vdcName = testConfig.VCD.Vdc
+	}
+	return func(s *terraform.State) error {
+		rs, ok := s.RootModule().Resources[node]
+		if !ok {
+			return fmt.Errorf("not found: %s", node)
+		}
+
+		if rs.Primary.ID == "" {
+			return fmt.Errorf("no VM ID is set")
+		}
+
+		conn := testAccProvider.Meta().(*VCDClient)
+		_, vdc, err := conn.GetOrgAndVdc(orgName, vdcName)
+		if err != nil {
+			return fmt.Errorf(errorRetrievingVdcFromOrg, vdcName, orgName, err)
+		}
+
+		_, err = vdc.QueryVmByName(vmName)
+		if err != nil {
+			return err
+		}
+
+		return nil
+	}
+}
+
+func testAccCheckVcdStandaloneVmDestroy(vmName string, orgName string, vdcName string) resource.TestCheckFunc {
+	if orgName == "" {
+		orgName = testConfig.VCD.Org
+	}
+	if vdcName == "" {
+		vdcName = testConfig.VCD.Vdc
+	}
+	return func(s *terraform.State) error {
+		conn := testAccProvider.Meta().(*VCDClient)
+
+		for _, rs := range s.RootModule().Resources {
+			if rs.Type != "vcd_vm" {
+				continue
+			}
+			_, vdc, err := conn.GetOrgAndVdc(orgName, vdcName)
+			if err != nil {
+				return fmt.Errorf(errorRetrievingVdcFromOrg, vdcName, orgName, err)
+			}
+
+			_, err = vdc.QueryVmByName(vmName)
+
+			if err == nil {
+				return fmt.Errorf("VM still exist")
+			}
+
+			return nil
+		}
+
+		return nil
+	}
+}
